SchemaIndex deadlock fix
Concurrent schema gets, that cannot already be served via the cache are being aggregated via a sync.Cond. The first go routine to fetch the schema is meant to send a broadcast message to all the other waiting go routines, such that they can retrieve the schema from the schema cache map. In certain error situation where the first goroutine would end with an error, the secondary goroutines would wait forever for the broadcast message. This PR fixes this issues