Fixing the deletion of keys
The deletion occures after importing the already committed intent from the cache in tree_proto format. The Intent content in the tree was marked as delete, as expected, but then re-applying the new intent content, the key attributes where just checked for existence, but the delete flag was not reset.